+++ Framework/Audio/Synthesizer/Effects/SoftClipperEffect.cs
@@ -1,0 +1,25 @@
+﻿namespace Macabre2D.Framework {
+
+    using System;
+    using System.Runtime.Serialization;
+
+    /// <summary>
+    /// Performs soft clipping on samples.
+    /// </summary>
+    /// <seealso cref="Macabre2D.Framework.Audio.Synthesizer.Effects.IAudioEffect"/>
+    [DataContract]
+    public sealed class SoftClipperEffect : IAudioEffect {
+
+        /// <summary>
+        /// Gets or sets the pre gain.
+        /// </summary>
+        /// <value>The pre gain.</value>
+        [DataMember]
+        public float PreGain { get; set; } = 1f;
+
+        /// <inheritdoc/>
+        public float ApplyEffect(float sample, float time) {
+            return (float)(2f / Math.PI) * (float)Math.Atan(this.PreGain * sample);
+        }
+    }
+}