+++ Framework/Audio/Synthesizer/Pitch.cs
@@ -1,0 +1,11 @@
+﻿namespace Macabre2D.Framework {
+
+    /// <summary>
+    /// Represents a pitch, which modifies a note up or down.
+    /// </summary>
+    public enum Pitch : byte {
+        Low = 3,
+        Normal = 4,
+        High = 5,
+    }
+}