+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 * MacabreGame object now has GameSpeed property which can dictate how fast time dependent components and modules are running.
+* CosmicSynth has been folded into the Macabre2D.Framework project.
 
 ### Refactors